refactor: more prop interface extraction and fixing imports
