+++ apps/web/src/components/Modals/ModalBase/ModalBody.tsx
@@ -1,5 +1,3 @@
-import React from "react"
-
 export default function ModalBody({
   children
 }: {
