+++ apps/web/src/components/layouts/AppLayout/Sidebar/Sidebar.tsx
@@ -1,14 +1,14 @@
 "use client"
 
+import { Fragment } from "react"
+import { motion } from "framer-motion"
+import Link from "next/link"
+import { LuX } from "react-icons/lu"
 import { User } from "@/app/context/AuthContext"
-import { Separator } from "@/components/Separator"
+import { Separator } from "@/components"
 import { cn } from "@mav/shared/utils"
 import { Button } from "@mav/ui/components/buttons"
 import { MyArtverseIcon } from "@mav/ui/icons"
-import { motion } from "framer-motion"
-import Link from "next/link"
-import { Fragment } from "react"
-import { LuX } from "react-icons/lu"
 import Menu from "../../../Menus"
 import { SidebarItems, fmDuration } from "./SidebarItems"
